+++ packages/cli/src/commands/org/resources/pull.spec.ts
@@ -7,8 +7,8 @@ jest.mock('../../../lib/snapshot/snapshotFactory');
 jest.mock('../../../lib/project/project');
 jest.mock('../../../lib/utils/process');
 
+import {join} from 'path';
 import {Config} from '../../../lib/config/config';
-import {mocked} from 'ts-jest/utils';
 import {ResourceSnapshotsReportType} from '@coveord/platform-client';
 import {test} from '@oclif/test';
 import {getDummySnapshotModel} from '../../../__stub__/resourceSnapshotsModel';
@@ -19,16 +19,19 @@ import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';
 import Command from '@oclif/command';
 import {IsGitInstalled} from '../../../lib/decorators/preconditions';
 import {PreconditionError} from '../../../lib/errors/preconditionError';
+import {cwd} from 'process';
+import {cli} from 'cli-ux';
 
-const mockedSnapshotFactory = mocked(SnapshotFactory, true);
-const mockedConfig = mocked(Config);
+const mockedSnapshotFactory = jest.mocked(SnapshotFactory, true);
+const mockedConfig = jest.mocked(Config);
 const mockedConfigGet = jest.fn();
 const mockedGetSnapshot = jest.fn();
 const mockedDownloadSnapshot = jest.fn();
 const mockedDeleteSnapshot = jest.fn();
-const mockedIsGitInstalled = mocked(IsGitInstalled, true);
-const mockedAuthenticatedClient = mocked(AuthenticatedClient);
+const mockedIsGitInstalled = jest.mocked(IsGitInstalled, true);
+const mockedAuthenticatedClient = jest.mocked(AuthenticatedClient);
 const mockEvaluate = jest.fn();
+const pathToStub = join(cwd(), 'src', '__stub__');
 
 const doMockConfig = () => {
   mockedConfigGet.mockReturnValue(
@@ -139,15 +142,16 @@ describe('org:resources:pull', () => {
     .stderr()
     .command(['org:resources:pull'])
     .it('should select all resource types', () => {
+      const resourcesToExport = {
+        FIELD: ['*'],
+        FEATURED_RESULT: ['*'],
+        SOURCE: ['*'],
+        QUERY_PIPELINE: ['*'],
+        SEARCH_PAGE: ['*'],
+        EXTENSION: ['*'],
+      };
       expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
-        expect.arrayContaining([
-          'FIELD',
-          'FEATURED_RESULT',
-          'SOURCE',
-          'QUERY_PIPELINE',
-          'SEARCH_PAGE',
-          'EXTENSION',
-        ]),
+        expect.objectContaining(resourcesToExport),
         'default-org',
         expect.objectContaining({})
       );
@@ -156,14 +160,116 @@ describe('org:resources:pull', () => {
   test
     .stdout()
     .stderr()
-    .command(['org:resources:pull', '-r', 'field', 'featuredResult', 'source'])
+    .command(['org:resources:pull', '-r', 'FIELD', 'FEATURED_RESULT', 'SOURCE'])
     .it('should select specified resource types', () => {
+      const resourcesToExport = {
+        FIELD: ['*'],
+        FEATURED_RESULT: ['*'],
+        SOURCE: ['*'],
+      };
       expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
-        ['FIELD', 'FEATURED_RESULT', 'SOURCE'],
+        resourcesToExport,
         'default-org',
         expect.objectContaining({})
       );
     });
+
+  test
+    .stdout()
+    .stderr()
+    .stub(cli, 'confirm', () => async () => true)
+    .command([
+      'org:resources:pull',
+      '-m',
+      join(pathToStub, 'snapshotPullModels', 'full.json'),
+    ])
+    .it(
+      'should create a snapshot with all resource types specified in the model',
+      () => {
+        const resourcesToExport = {
+          FIELD: ['*'],
+          QUERY_PIPELINE: ['*'],
+          ML_MODEL: ['*'],
+          SUBSCRIPTION: ['*'],
+        };
+        expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
+          resourcesToExport,
+          expect.anything(),
+          expect.objectContaining({})
+        );
+      }
+    );
+
+  test
+    .stdout()
+    .stderr()
+    .stub(cli, 'confirm', () => async () => true)
+    .command([
+      'org:resources:pull',
+      '-m',
+      join(pathToStub, 'snapshotPullModels', 'subset.json'),
+    ])
+    .it(
+      'should create a snapshot with only the specified resource items in the model',
+      () => {
+        const resourcesToExport = {
+          FIELD: ['author', 'source', 'title'],
+          QUERY_PIPELINE: ['default', 'agentPanel'],
+          ML_MODEL: null,
+          SEARCH_PAGE: null,
+          SUBSCRIPTION: null,
+        };
+        expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
+          resourcesToExport,
+          expect.anything(),
+          expect.objectContaining({})
+        );
+      }
+    );
+
+  test
+    .stdout()
+    .stderr()
+    .stub(cli, 'confirm', () => async () => true)
+    .command([
+      'org:resources:pull',
+      '-m',
+      join(pathToStub, 'snapshotPullModels', 'full.json'),
+    ])
+    .it('should use the orgId from the model', () => {
+      expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
+        expect.anything(),
+        'myorgid',
+        expect.objectContaining({})
+      );
+    });
+
+  test
+    .stdout()
+    .stderr()
+    .command([
+      'org:resources:pull',
+      '-m',
+      join(pathToStub, 'snapshotPullModels', 'missingOrgId.json'),
+    ])
+    .it('should use the orgId from the config', () => {
+      expect(mockedSnapshotFactory.createFromOrg).toHaveBeenCalledWith(
+        expect.anything(),
+        'default-org',
+        expect.objectContaining({})
+      );
+    });
+
+  test
+    .stdout()
+    .stderr()
+    .command([
+      'org:resources:pull',
+      '-m',
+      join(pathToStub, 'snapshotPullModels', 'missingResources.json'),
+    ])
+    .catch(/requires property "resourcesToExport"/)
+    .it('should throw an invalid model error');
 
   test
     .stdout()