+++ packages/cli/src/lib/config/configRenderer.spec.ts
@@ -3,7 +3,6 @@ jest.mock('../../lib/config/config');
 import {BaseConfiguration, Config} from './config';
 import {fancyIt} from '../../__test__/it';
 import {ConfigRenderer} from './configRenderer';
-import {mocked} from 'ts-jest/utils';
 
 describe('configRenderer', () => {
   const userFacingConfigKeys: (keyof BaseConfiguration)[] = [
@@ -17,7 +16,7 @@ describe('configRenderer', () => {
   } as unknown as Config;
 
   const doMockConfig = () => {
-    const mockedConfig = mocked(Config);
+    const mockedConfig = jest.mocked(Config);
     mockedConfig.userFacingConfigKeys = userFacingConfigKeys;
   };