+++ packages/cli/src/lib/decorators/preconditions/npx.spec.ts
@@ -1,7 +1,6 @@
 jest.mock('../../utils/process');
 
 import {dedent} from 'ts-dedent';
-import {mocked} from 'ts-jest/utils';
 import {spawnProcessOutput} from '../../utils/process';
 import {getFakeCommand} from './testsUtils/utils';
 
@@ -11,7 +10,7 @@ import {fancyIt} from '../../../__test__/it';
 import {PreconditionError} from '../../errors/preconditionError';
 
 describe('IsNpxInstalled', () => {
-  const mockedSpawnProcessOutput = mocked(spawnProcessOutput);
+  const mockedSpawnProcessOutput = jest.mocked(spawnProcessOutput);
   beforeEach(() => {
     jest.resetAllMocks();
   });