+++ packages/cli/src/lib/platform/url.spec.ts
@@ -1,13 +1,16 @@
 import {Region} from '@coveord/platform-client';
-import {mocked} from 'ts-jest/utils';
 import {fancyIt} from '../../__test__/it';
 
 jest.mock('./environment');
 import {PlatformEnvironment, platformUrl} from './environment';
-import {snapshotSynchronizationUrl, snapshotUrl} from './url';
+import {
+  createSnapshotUrl,
+  snapshotSynchronizationUrl,
+  snapshotUrl,
+} from './url';
 
 describe('url', () => {
-  const mockedPlatformUrl = mocked(platformUrl);
+  const mockedPlatformUrl = jest.mocked(platformUrl);
   beforeEach(() => {
     mockedPlatformUrl.mockReturnValue('https://foo.test');
   });
@@ -43,6 +46,24 @@ describe('url', () => {
         })
       ).toBe(
         'https://foo.test/admin/#some-org/organization/resource-snapshots/some-snapshot/synchronization'
+      );
+
+      expect(platformUrl).toBeCalledWith({
+        environment: PlatformEnvironment.QA,
+        region: Region.AU,
+      });
+    });
+  });
+
+  describe('#createSnapshotUrl', () => {
+    fancyIt()('should build the URL properly', () => {
+      expect(
+        createSnapshotUrl('some-org', {
+          environment: PlatformEnvironment.QA,
+          region: Region.AU,
+        })
+      ).toBe(
+        'https://foo.test/admin/#some-org/organization/resource-snapshots/create-snapshot'
       );
 
       expect(platformUrl).toBeCalledWith({